+++ web/mazes/recursive_division.js
@@ -1,11 +1,18 @@
+// @flow
 import _ from 'underscore'
 
+import { Grid } from './grid.js'
+
 export default class RecursiveDivision {
+  grid: Grid
+  maxRoomSize: number
+  roomPct: number
+
   static on(grid) {
     (new RecursiveDivision(grid)).run()
   }
 
-  constructor(grid) {
+  constructor(grid: Grid) {
     this.grid = grid
     this.maxRoomSize = _.random(5)
     this.roomPct = _.random(6)
@@ -23,7 +30,7 @@ export default class RecursiveDivision {
     this.divide(0, 0, this.grid.rows, this.grid.cols)
   }
 
-  divide(row, col, height, width) {
+  divide(row: number, col: number, height: number, width: number) {
     if (height <= 1 || width <= 1 || this.makeRoom(height, width))
       return
 
@@ -34,11 +41,11 @@ export default class RecursiveDivision {
     }
   }
 
-  makeRoom(height, width) {
+  makeRoom(height: number, width: number) {
     return height < this.maxRoomSize && width < this.maxRoomSize && _.random(this.roomPct) == 0
   }
 
-  horizontal(row, col, height, width) {
+  horizontal(row: number, col: number, height: number, width: number) {
     let divideSouthOf = _.random(height-2)
     let passageAt = _.random(width-1)
 
@@ -46,14 +53,15 @@ export default class RecursiveDivision {
       if (passageAt == x)
         continue
       let cell = this.grid.get(row+divideSouthOf, col+x)
-      cell.unlink(cell.south)
+      if (cell)
+        cell.unlink(cell.south)
     }
 
     this.divide(row, col, divideSouthOf + 1, width)
     this.divide(row + divideSouthOf + 1, col, height - divideSouthOf - 1, width)
   }
 
-  vertical(row, col, height, width) {
+  vertical(row: number, col: number, height: number, width: number) {
     let divideEastOf = _.random(width-2)
     let passageAt = _.random(height-1)
 
@@ -61,7 +69,8 @@ export default class RecursiveDivision {
       if (passageAt == y)
         continue
       let cell = this.grid.get(row+y, col+divideEastOf)
-      cell.unlink(cell.east)
+      if (cell)
+        cell.unlink(cell.east)
     }
 
     this.divide(row, col, height, divideEastOf + 1)
